feat(sidebars-mutations): moved sidebar component into single place
