+++ examples/shop-with-blog/client/src/components/Sidebar/index.js
@@ -1,2 +1,4 @@
 export * from './CloseSidebarMutation';
 export * from './OpenSidebarMutation';
+export * from './SidebarQuery';
+export * from './SidebarContainer';
