--- examples/shop-with-blog/client/src/pages/shop/components/Sidebar/index.js
@@ -1,2 +1,0 @@
-export * from './SidebarContainer';
-export * from './SidebarQuery';
